Dynamic Rule model, validation and Verbalization
Added model elements, validation and verbalization for dynamic rules
(aka event rules). The model change introduced an abstract GeneralRule
with an editable name (validated as unique in the model) but no rule
implementation. A DynamicRule is a subtype of GeneralRule that owns
a single LeadRolePath that allows markup for dynamic states.

Other RolePathOwner implementations allow significantly more use of
top-level objects (multiple root paths, shared paths and subqueries).
Due to the extremely rule-specific markup required in the path to
implement a dynamic rule these sharing and reuse notions are not
supported with a dynamic rule. This also required virtualization of
the RolePathOwner concept as IRolePathOwner since DynamicRule is a
subtype of GeneralRule, not RolePathOwner.

A dynamic rule is implemented as a static rule that can apply to
multiple object states: initial is the starting state for a model
change, added represents new elements and deleted represents elements
removed from the initial state. The current state (where all other
rules are assumed to exist) is a union of the initial and added states
without the deleted state. To reason about these states a rule has
4 logical parts: the context quantifier defines variables that are
used in the head, the rule head adds to objects and facts, and the
rule body provides the conditions.

Getting these 4 sections coordinated is a difficult exercise, and role
paths always automatically create variables anyway when translated to
first-order predicate logic. Therefore, instead of forcing the user to
manage all of these sections we distinguish between elements added or
deleted in the head (actions taken by the rule) and added or deleted
elements in the body. With this approach, we can determine variables
and placement by defining two more states: add (instead of added)
refers to objects and facts created in the head. Similarly for delete.

With these two new states, added and deleted are treated as events and
add and delete as actions. While logically the states are the same,
this makes the role path definition easier to conceptualize because
the placement (declarations, head add section, head delete section and
body) can be determined from the specified node state. To verbalize
this approach, any variable that is referenced by an add or delete
fact, along with quantifiers for 'delete' objects are added to the
head variable list along with state annotation. The 'add' variables
(representing new instances) are verbalized in 'Add:' section along
with a list of 'add' fact types.

The 'add' and 'delete' elements are extracted from the role path
verbalization, which is only shown if there are remaining elements.
Dynamic state like '(fact type) is added' is also added to the path
verbalization.