Verbalization Report Fails with Simple Model
The vast majority of models have some non-reported object types and
fact types. These occur with objectification, spanning uniqueness
constraints (which cause implicit objectification) and unary fact
types. There is an original defect in the generator that tries to sort
an uneditable list if nothing is filtered and the report generator
fails silently for trivial models.