ORMCustomTool failing if .orm file opened as text
The custom tool is failing if the .orm file is opened in a text
editor in Visual Studio, including xml editors. The tool is asking
the document for specific automation objects that these editors
do not recognized. This is harmless and fully recoverable, but only
when the custom tool swallows the exception.