UUID data type unnecessarily forces extra table
Use of the UUID data type with the default auto generated setting
(DataTypeAutoGenerated = true) blocks the identified entity type from
being absorbed during relational mapping. This is intentional behavior
for auto generated data types and is appropriate when the identifier
is generated incrementally. However, the UUID data type generation
algorithm uses a default column bound to a function.

There is no need to block absorption with a function-bound generator,
so this change introduces this extra condition.

This is tied to an algorithm version change at the OIAL level, which
generally triggers a full rebuild on file load. However, this version
transition is only applied if the UUIDNumericDataType is used with
auto generation enabled, so there will be no on-load change for files
that do not use this construct.