Unary negation, isSUBTYPE columns, default values
Major changes for July 2024 NORMA release.

This completely eliminates unary binarization in NORMA and replaces it
with automatically managed unary negation fact types. This allows much
tighter control over the allowed values, mandatory states on negatable
unaries, default true/false values, objectification and constraining
of the negated state. New notations are also introduced to indicate
when a unary fact type is paired with its inverse.

isSUBTYPE columns determine if an absorbed object type (subtype, unary
objectification, etc.) can be recognized as a subtype base on other
evidence in the model (essentially columns mapping to mandatory roles).
If this is not found, a new isSUBTYPE column is added to the absorbing
table.

Default values allow DDL default value settings. These cannot be
mapped to absorbed subtypes, so the modeler should use a separate
table to get default values on subtypes.