Allow Role.DefaultValue property redirection
The property descriptor for the role DefaultValue property is being
redirected from a relational column. However, the associated
singleton TypeConverter only knows about Role and ObjectType, not
Column, so the Column.DefaultValue property just shows System.Object.

Use an instance-bound converter instead of a static one so we do not
need to use context.Instance to resolve the object being displayed.