ORMCustomTool incompatible project message
The ORMCustomTool generator currently only supports the classic
project file, which is an msbuild file format. The previous failure in
this case was just a generic null reference exception.

We know quite early if this fails because a root project item cannot
be found. Provide an exception message (or error in the generated C#
file) if the project cannot be resolved as the expected type.

The ideal solution is obviously to support the project type, but this
is a large work item touching at least 3000 lines of code. For now, at
least let the user know what the problem is.