Consider derivation state for isSUBTYPE columns
If an absorbed subtype is fully derived and not stored then it should
not get an absorption marker (aka isSUBTYPE) column in the relational
mapping.

Similarly, an absorbed objectification should not get the extra
absorption marker if the objectified fact type is fully derived.

A fallback was added for the unusual edge case where a unary negation
with no reading clears the 'mappable' filter but the positive unary
does not. This left us without a candidate for a column name,
resulting in COLUMN as the name. It now prepends a 'not' to the the
positive reading.