Table not created for type with only unary roles
Unary roles were missed when determining if an entity type plays roles
outside the preferred identifier. This eventually led to no table in
simple cases where the entity type did not play non-unary roles and
was not forcibly created for another reason (such as an auto-generated
identifier).

A unary role may also be implicitly mandatory if it is the only role
on a non-independent entity type. We normally do not allow mandatory
unary roles (if the unary negation is defined, then the mandatory is
disjunctive), but if this is the only role then it is implicitly
mandatory. Although this is very much an edge case, we still need to
map it accurately, so we check for the implied mandatory constraint.