Check implied mandatory for paired optional unary
The UnaryPattern determines the mandatory constraint pattern on a
negatable unary, so deeper constraint analysis is not usually needed.
The one edge case here is an implied mandatory constraint, which is
controlled by the object type, not the unary fact type.

If you have one TrueFalseUnspecified unary on an object type and
nothing else then the paired unary should also be considered pairwise
mandatory. This has to be checked in addition to the unary pattern
for this minimal edge case.

'A(.Id) r~' in the fact editor produces all of the model you need to
test this case.