Add error states for SubqueryParameterInput
Added errors when not all parameters are associated with elements in
the calling path, and when an input is bound but bound item has a
type mismatch with the parameter type.

I also added a subquery parameter input as an acceptable use of a
calculation and adjusted the error message accordingly.

Subquery inputs were not specified as known relationships in the
copy/merge closure, so the derived fact type was missing all inputs
when dropped in a different model.

Note that there are still parts of the subquery parameter area that
do not have error validation. This change covers the input binding,
not the QueryParameter and QueryParameterBinding objects used to
define the subquery and attach parameters to path nodes.