Model Browser Reference Node EditLabel Command
The ORM Model Browser checks if a survey node name is editable or if
it supports custom editing on the survey node. However, if the primary
node is not a survey node and the referenced element is then we also
need to check for survey node support on the reference.

This left a strange situation where a survey node would auto activate
an in place editor when selected but could not be reactivated with
an F2. Inline dropdowns (since VS2019) are opened with a SendKeys
F2 then Alt-Down. The F2 failed, and the alt was ignored, so the
selection just move down a line in this case.

Note that this is a safe fix because this is only step one in
determining if an inplace editor can be activated and/or a dropdown
opened. There is an extensive handshake with ISurveyNodeCustomEditor
that runs before this can be applied. Allowing the command through
just lets the more sophisticated code execute.