Tighten Unary Pattern Editing Gestures
There were a few loopholes in the editor that allowed users to modify
elements of the constraints implied by unary negation. Modifying or
deleting these constraints (implied internal uniqueness, disjunctive
mandatory constraint, exclusion constraint) is now blocked from all
editor attacks. Attempting to edit these constraints on a diagram will
just give a message telling the user that the constraints are not
editable.

In addition, there were two patterns that could be established that
mimicked the unary pattern constraints but were not reflected in the
pattern or mapping. Generally, if a constraint is added that mirrors
a pattern-controlled constraint then it will produce a visible error.
There were two cases where this did not happen:

1) A disjunctive mandatory constraint is manually added to each role
in the unary negation pair. If the UnaryPattern is not Required, then
there is no constraint to contradict, so this sits there but is not
part of the pattern, does not affect relational mapping, etc.
2) If the pattern is True/False obligatory (deontic mandatory in the
pattern) and an alethic disjunction is manually applied to the two
roles, then the implication error is in the pattern mandatory.

In the first case, we simply incorporate the mandatory constraint into
the negation and adjust the pattern accordingly. The user will see the
pattern glyphs change on the unary, the disjunctive mandatory change
to an XOr, and any shapes for the controlled exclusion will disappear.

In the second case, the alethic mandatory replaces the deontic one in
the pattern and the deontic one is deleted.

If the assimilated constraint was original an xor, the exclusion is
the same as the one we already have, so it is deleted.