Drag unary negation constraint shapes
If both the positive and negative ends of a unary negation are visible
on a diagram, selecting the context menu for either fact type will now
show 'Drag Unary Negation constraint(s)'. This will put one constraint
on the drag cursor unless you are using the obligatory pattern, which
will give you two (the first click places the exclusion, the second
the deontic mandatory).

Note that this behaves the same as dragging constraints from the model
browser, which does not create a new shape for a constraint unless it
can attach all of its roles. This means that it is not blocked from
attaching from another closer shape for the same constraint. The only
check made to show the command is that the diagram has at least one
connectable shape (DisplayAsObjectType blocks connection) for each
paired fact type. There is no way to predict if a drop location will
succeed.