Improve unary default value presentation
The default value for a unary fact type is stored as True internally,
but that is just a placeholder indicating that the role is populated.
It is not worth changing the file format by moving the unary
population to DefaultState instead of a DefaultValue, but it is worth
presenting this differently to the user.

A unary role is not a value role, so it technically does not have a
default value. However, the role has a default population. Saying the
role is populated also indicates the fact is populated since there is
only one role.

The DefaultValue property has kept its name for consistency with other
roles, but now says <Not Populated by Default> or (in bold, indicating
a non-default state) <Populated by Default> instead of being empty or
True. The default value is still controlled by the UnaryPattern, which
is why it is read only.

Instead of saying Default: True for the role, the verbalizer now says:
Each new instance of ROLEPLAYER plays this role by default.

This is a cosmetic change that is intended to add clarity to default
states available through the UnaryPattern property.